% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server.R
\name{check_server}
\alias{check_server}
\title{Check climatic data server}
\usage{
check_server(
  climatic_var = NULL,
  year = NULL,
  version = "last",
  verbose = TRUE,
  time_unit = "day"
)
}
\arguments{
\item{climatic_var}{Optional. One of "Prcp", "Tmin", or "Tmax".}

\item{year}{Optional. Year between 1950 and the latest year available on the
server. Use \code{get_periods()} to obtain the available data versions and years available.}

\item{version}{Character. Climatic data version. Either "last" (default) or
"4" (version 4).}

\item{verbose}{Logical. If TRUE, prints diagnostic messages; otherwise,
returns only TRUE/FALSE.}

\item{time_unit}{Character. One of "day" , "month" or "year".}
}
\value{
TRUE if the server is available; FALSE otherwise.
}
\description{
Checks that the online climatic data server is available and working correctly.
}
\examples{
interactive()
check_server()
}
