% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyclimate-package.R
\docType{package}
\name{easyclimate-package}
\alias{easyclimate}
\alias{easyclimate-package}
\title{easyclimate: Easy Access to High-Resolution Daily Climate Data for Europe}
\description{
Get high-resolution (1 km) daily, monthly and annual climate data (precipitation, and average, minimum and maximum temperatures) for points and polygons within Europe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://verughub.github.io/easyclimate}
  \item \url{https://github.com/VeruGHub/easyclimate}
  \item Report bugs at \url{https://github.com/VeruGHub/easyclimate/issues}
}

}
\author{
\strong{Maintainer}: Sofía Miguel \email{sofia.miromero@gmail.com} (\href{https://orcid.org/0000-0002-1341-8022}{ORCID})

Authors:
\itemize{
  \item Verónica Cruz-Alonso \email{veronica.cral@gmail.com} (\href{https://orcid.org/0000-0002-0642-036X}{ORCID}) [copyright holder]
  \item Francisco Rodriguez-Sanchez (\href{https://orcid.org/0000-0002-7981-1599}{ORCID}) [copyright holder]
  \item Christoph Pucher (\href{https://orcid.org/0000-0002-9269-1907}{ORCID})
  \item Paloma Ruiz-Benito (\href{https://orcid.org/0000-0002-2781-5870}{ORCID})
  \item Julen Astigarraga (\href{https://orcid.org/0000-0001-9520-3713}{ORCID})
  \item Mathias Neumann (\href{https://orcid.org/0000-0003-2472-943X}{ORCID})
  \item Sophia Ratcliffe (\href{https://orcid.org/0000-0001-9284-7900}{ORCID})
}

}
\keyword{internal}
