% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periods.R
\name{get_periods}
\alias{get_periods}
\title{Get available climatic data versions and years available.}
\usage{
get_periods()
}
\value{
A data.frame containing the available climate data versions and
their corresponding initial and latest available years.
}
\description{
Returns a data.frame describing the available climatic data versions and
their temporal coverage (initial and latest available years).
}
\references{
For details on the latest version of the climatic data, see:
Pucher, Christoph (2026). Description of Downscaled European Climate Data. figshare.
Online resource. https://doi.org/10.6084/m9.figshare.33078053.v1

For details on version 4, see:
Pucher C. 2023. Description and Evaluation of Downscaled Daily Climate Data Version 4.
https://doi.org/10.6084/m9.figshare.22962671.v1

Adam Moreno, Hubert Hasenauer. 2016. Spatial downscaling of European climate data.
International Journal of Climatology 36: 1444–1458.
}
\author{
Sofia Miguel
}
