% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.R
\name{build_url}
\alias{build_url}
\title{Build the URL for a given request}
\usage{
build_url(climatic_var_single, year, version = "last", temp_res = "day")
}
\arguments{
\item{climatic_var_single}{Character. Climatic variable to download.
One of "Tmax","Tmin", "Tavg", or "Prcp".}

\item{year}{Numeric. Year for which climatic data will be downloaded.}

\item{version}{Character. Climate data version. Either "last" (default) or
"4" (version 4).}

\item{temp_res}{Character. One of "day" , "month" or "year".}
}
\value{
Character string with the URL.
}
\description{
Build the URL to download climate data from
ftp://palantir.boku.ac.at/Public/ClimateData/
}
\author{
Veronica Cruz-Alonso, Francisco Rodríguez-Sánchez, Sophia Ratcliffe, Sofía Miguel
}
