% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.R
\name{build_url}
\alias{build_url}
\title{Build url address for a given request}
\usage{
build_url(climatic_var, year)
}
\arguments{
\item{climatic_var}{Character. Climatic variable to be downloaded. One of 'Tmax', 'Tmin' or 'Prcp'.}

\item{year}{Numeric. Year to download climatic information}
}
\value{
text string with the url
}
\description{
Build the url to download climatic data from ftp://palantir.boku.ac.at/Public/ClimateData/
}
\author{
Veronica Cruz-Alonso, Sophia Ratcliffe, Francisco Rodríguez-Sánchez
}
