% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server.R
\name{check_server}
\alias{check_server}
\title{Check climate data server}
\usage{
check_server(
  climatic_var = NULL,
  year = NULL,
  version = "last",
  verbose = TRUE,
  time_unit = "day"
)
}
\arguments{
\item{climatic_var}{Optional. One of "Prcp", "Tmin", or "Tmax".}

\item{year}{Optional. Year between 1950 and the latest year available.}

\item{version}{Character. Version of the climate data. Either 'last' or "4" version.}

\item{verbose}{Logical. Print diagnostic messages, or just return TRUE/FALSE?}

\item{time_unit}{Character. One of "Day", "Month", or "Year".}
}
\value{
TRUE if the server seems available, FALSE otherwise.
}
\description{
Check that the online climate data server is available and working correctly.
Only works for version 4 of the climate dataset.
}
\details{
This function checks access to the latest version of the climatic
dataset (version 4).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
check_server()
\dontshow{\}) # examplesIf}
}
