% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_server.R
\name{check_server}
\alias{check_server}
\title{Check climate data server}
\usage{
check_server()
}
\value{
TRUE if the server seems available, FALSE otherwise.
}
\description{
Check that the online climate data server is available and working correctly.
}
\examples{

check_server()

}
