% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyclimate-package.R
\docType{package}
\name{easyclimate-package}
\alias{easyclimate}
\alias{easyclimate-package}
\title{easyclimate: easy download and tidy up of a 1-km and 1-day resolution database of European climate}
\description{
Facilitate the use of the European climatic database 
    (ftp://palantir.boku.ac.at/Public/ClimateData/) containing daily 
    high (1 km) resolution data for precipitation, maximum and minimum temperature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VeruGHub/easyclimate}
  \item Report bugs at \url{https://github.com/VeruGHub/easyclimate/issues}
}

}
\author{
\strong{Maintainer}: Verónica Cruz-Alonso \email{veronica.cral@gmail.com} (\href{https://orcid.org/0000-0002-0642-036X}{ORCID})

Authors:
\itemize{
  \item Sophia Ratcliffe (\href{https://orcid.org/0000-0001-9284-7900}{ORCID})
  \item Francisco Rodriguez-Sanchez (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})
  \item Paloma Ruiz-Benito (\href{https://orcid.org/0000-0002-2781-5870}{ORCID})
  \item Julen Astigarraga (\href{https://orcid.org/0000-0001-9520-3713}{ORCID})
}

}
\keyword{internal}
