% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_climate.R
\name{get_daily_climate}
\alias{get_daily_climate}
\title{Get daily climate data}
\usage{
get_daily_climate(
  coords = NULL,
  climatic_var = "Prcp",
  period = NULL,
  output = "df",
  ...
)
}
\arguments{
\item{coords}{A matrix, dataframe, \code{\link[sf:sf]{sf::sf()}}, or \code{\link[terra:SpatVector-class]{terra::SpatVector()}} object
containing point or polygon coordinates (in lonlat/geographic format).
If \code{coords} is a matrix, it must have only two columns: the first with longitude
and the second with latitude data.
If \code{coords} is a data.frame, it must contain two columns called \code{lon} and \code{lat}
with longitude and latitude coordinates, respectively.}

\item{climatic_var}{Character. Climatic variable to be downloaded. One of 'Tmax', 'Tmin' or 'Prcp'.}

\item{period}{Either a single number (representing a year between 1950 and 2017),
a date in "YYYY-MM-DD" format (to obtain data for a single day),
or a vector of length 2 specifying the starting and ending dates.}

\item{output}{Character. Either "df", which returns a dataframe with daily climatic values
for each point/polygon, or "raster", which returns a \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} object.}

\item{...}{further arguments for \code{\link[terra:extract]{terra::extract()}}. Note you could use this to
directly calculate summary statistics (e.g. mean) for each polygon.}
}
\value{
A data.frame or a \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} object (if output = "raster").
}
\description{
Extract daily climate data for a given set of points or polygons.
}
\examples{
\dontrun{
library(easyclimate)

# Coords as matrix
coords <- matrix(c(-5.36, 37.40), ncol = 2)
ex <- get_daily_climate(coords, period = "2008-09-27")  # single day
ex <- get_daily_climate(coords, period = 2008)  # entire year
ex <- get_daily_climate(coords, period = c(2008, 2010))  # several years
ex <- get_daily_climate(coords, period = c("2008-09-27", "2008-09-30"))  # specific period

ex <- get_daily_climate(coords, period = "2008-09-27", climatic_var = "Tmin")

# Coords as data.frame
coords <- as.data.frame(coords)
names(coords) <- c("lon", "lat")  # must have these columns
ex <- get_daily_climate(coords, period = "2008-09-27")  # single day

# Coords as sf
coords <- sf::st_as_sf(coords, coords = c("lon", "lat"))
ex <- get_daily_climate(coords, period = "2008-09-27")  # single day

# Several points
coords <- matrix(c(-5.36, 37.40, -4.05, 38.10), ncol = 2, byrow = TRUE)
ex <- get_daily_climate(coords, period = "2008-09-27", output = "raster")  # raster output

## Polygons
coords <- terra::vect("POLYGON ((-5 38, -5 37.5, -4.5 37.5, -4.5 38, -5 38))")

# Return raster
ex <- get_daily_climate(coords, period = "2008-09-27", output = "raster")

# Calculate average across polygon
ex <- get_daily_climate(coords, period = "2008-09-27", fun = "mean")
# Calculate min across polygon
ex <- get_daily_climate(coords, period = "2008-09-27", fun = "min")


# easily convert sf to SpatVector
# coords <- vect(poly.sf)
}

}
\author{
Veronica Cruz-Alonso, Sophia Ratcliffe, Francisco Rodríguez-Sánchez
}
